MP-202 :truck: Rename: servlet 폴더 이름을 http로 변경
* 프로젝트 컨벤션 준수
