+++ src/main/java/kr/modusplant/domains/communication/conversation/app/controller/ConvCommentController.java
@@ -5,7 +5,7 @@ import kr.modusplant.domains.communication.conversation.app.http.response.ConvCo
 import kr.modusplant.domains.communication.conversation.app.service.ConvCommentApplicationService;
 import kr.modusplant.domains.communication.conversation.persistence.entity.ConvPostEntity;
 import kr.modusplant.domains.member.persistence.entity.SiteMemberEntity;
-import kr.modusplant.global.app.servlet.response.DataResponse;
+import kr.modusplant.global.app.http.response.DataResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.http.ResponseEntity;
