+++ src/main/java/kr/modusplant/domains/communication/qna/app/controller/QnaPostController.java
@@ -8,7 +8,7 @@ import kr.modusplant.domains.communication.qna.app.http.request.QnaPostInsertReq
 import kr.modusplant.domains.communication.qna.app.http.request.QnaPostUpdateRequest;
 import kr.modusplant.domains.communication.qna.app.http.response.QnaPostResponse;
 import kr.modusplant.domains.communication.qna.app.service.QnaPostApplicationService;
-import kr.modusplant.global.app.servlet.response.DataResponse;
+import kr.modusplant.global.app.http.response.DataResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Pageable;
