+++ src/main/java/kr/modusplant/domains/communication/tip/app/controller/TipCategoryController.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.modusplant.domains.communication.tip.app.http.request.TipCategoryInsertRequest;
 import kr.modusplant.domains.communication.tip.app.http.response.TipCategoryResponse;
 import kr.modusplant.domains.communication.tip.app.service.TipCategoryApplicationService;
-import kr.modusplant.global.app.servlet.response.DataResponse;
+import kr.modusplant.global.app.http.response.DataResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.http.ResponseEntity;
