+++ src/main/java/kr/modusplant/modules/auth/normal/app/controller/NormalSignUpController.java
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import kr.modusplant.domains.term.app.http.response.TermResponse;
-import kr.modusplant.global.app.servlet.response.DataResponse;
+import kr.modusplant.global.app.http.response.DataResponse;
 import kr.modusplant.modules.auth.normal.app.http.request.NormalSignUpRequest;
 import kr.modusplant.modules.auth.normal.app.service.NormalSignUpApplicationService;
 import lombok.RequiredArgsConstructor;
